Rename trade.go to DONE_trade.go
This was very helpful for my understanding of the rational for using the mutex's throughout the library. I see now that it is being used to manage the flow of information from TWS to the trading application. These channels are opened and closed as needed based on the current trades/other processes currently exchanging information back and forth. IE, a trade will open a channel for the flow of information as it relates to THAT trade, going forward all the trades information will pass through this until such a time that the trade is finished and then the "done()" method is called, which closes that chennel. Therefore closing the communication stream back and forth for that trade.

Additionally, I really love this implementation of "trade", the initial struct being built with "fills" and "tradeLog" is great. Plus you have access to lots of information using "trade" and its methods. I will be relying on this for sure. Lastly, look into utilizing the logs and trade details to build a comprehensive history of my trades.